Implement purging on either of hash or url.
In VCL:
	purge_url(<regexp>)
	purge_hash(<regexp>)
	(for an interrim period purge(<regexp>) will be the same as purge_url).
In CLI
	url.purge <regexp>
	hash.purge <regexp>

purge_hash operates on the hash-string which results from vcl_hash(),
by default it is composed of:

	req.url "#" req.http.host "#"

To purge everything on the virtual host foo.bar.com:

In CLI:
	url.purge "#foo.bar.com#$"

In VCL:
	purge_hash("#foo.bar.com#$");


The general format, if you have defined vcl_hash(), is:

	Each "req.hash +=" operator appends the right hand side of the
	+= and a "#" separator.

You'll have to figure out your own regexps.


Under the hood:

Move the hash string from object to objecthead and save space while
we're at it.

Fix indentation in generated source.



git-svn-id: http://www.varnish-cache.org/svn/trunk/varnish-cache@1816 d4fa192b-c00b-0410-8231-f00ffab90ce4