Add a toplevel word which examines the sequence returned by srandom(1) and stops the test if we do not get the same sequence as we expect.
The Open Group does not define which deterministic sequence srandom(1)
should result in, on that it be deterministic, but I have high hopes
in the general sanity and expect that UNIX people across the board
have realized that for portability the same sequence should be
returned on all platforms.

At the very least FreeBSD and Linux/GLIBC, as seen on projects.linpro.no,
agree.



git-svn-id: http://www.varnish-cache.org/svn/trunk/varnish-cache@3367 d4fa192b-c00b-0410-8231-f00ffab90ce4