Continue the civilization of varnishtest:
Unify the construction of the workdir name.

Append uid to the name if non-zero (as in: "/tmp/__v1.488") to avoid
multiple users stomping on each other.

Blow away and recreate the workdir at the creation of an instance.

Close and reopen the shmem for each launch, close during cleanup.

Fixes #13



git-svn-id: http://www.varnish-cache.org/svn/trunk/varnish-cache@3976 d4fa192b-c00b-0410-8231-f00ffab90ce4