On Solaris we occasionally get EBADF when we use the ioctl to toggle blocking. This then results in the connects failing.
The way Solaris seems to want it to be done is using fcntl
over two syscalls. So make it so conditionally for Solaris.

Should fix #482 and #535

Compiled on Solaris and Victori tested a prelminary version
of the patch that solved it. Sadly test cases don't 
seem to run at all on my solaris dev zone.



git-svn-id: http://www.varnish-cache.org/svn/trunk/varnish-cache@4161 d4fa192b-c00b-0410-8231-f00ffab90ce4