Move the varnishstat -f argument into varnishapi, it is broadly applicable and can be implemented more efficiently this way.
The new syntax is an expansion of the old, which allows
"class.ident.name" or just "name" patterns to be specified.

A pattern which starts with '^' excludes the matching fields,
otherwise the pattern includes them.

If the first pattern is an exclude pattern, all fields start
out included, and vice versa.

Notice that "class" and "ident" fields always must have their
trailing '.' even if they are empty.

Each subcomponent can have a '*' suffix, which will ("glob") match
anything.

Multiple "-f" arguments can be given, they will be concatenated.

All patterns will be tested sequentially, and the last one that
matches determines the result.

Example:

	-f "SMA..sma_nreq,^SMA.ba*." -f "shm_*"

This will return:
	SMA.storage_0.sma_nreq
	shm_records
	shm_writes
	...

but not:
	SMA.storage_0.sma_nobj
	SMA.bar.sma_nreq
	sms_nreq
	...




git-svn-id: http://www.varnish-cache.org/svn/trunk/varnish-cache@4911 d4fa192b-c00b-0410-8231-f00ffab90ce4