Fix a long-standing bug in varnishtest:
It used to be that we'd call http processing on a filedescriptor and
then that was that.  Then we added keywords to do new accepts in the
server and suddenly the calling code closes a wrong filedesc which
belongs to somebody else and things get really inconvenient fast.

This made all test-cases which use the "accept" server directive flakey.