Add a new mode to server instance 'b0', where it acts like "a normal webserver" and dispatches a thread for each incomming connection.
This is an example use:

	server s0 {
		loop 10 {
			rxreq
			txresp -body "foo1"
		}
		rxreq
		txresp -hdr "Connection: close" -body "foo1"
		expect_close
	} -dispatch

Each connection will spawn a dynamically created s%d instance
starting with s1, s2 ...

Each of these will respond to 11 requests on the accepted connection,
the last response will have "Connection: close" and they will
expect the other end (varnish) to do that.

The main limitation on using this for bulk traffic tests is the
finite and limited size of the vtc_log which varnishtest will
collect (half a megabyte).

Test b00048 is about as much traffic as will fit in the vtc_log.