Use closefd() throughout, take 2
This patch was created using Coccinelle and the following steps:

    $ cat >closefd.cocci <<EOF
    @@
    expression fd;
    @@

    - AZ(close(fd));
    + closefd(&fd);
    EOF
    $ spatch --dir . --in-place --sp-file closefd.cocci

Please note that one use of AZ(close(...)) couldn't be moved to
closefd() because of a const qualifier. There are other calls to
close(2) that are currently not checked on purpose. They may need
to be revisited.

See: git grep -F '(void)close'