Introduce VCL_STRANDS, and use it for string compare.
VCL_STRANDS is an alternative to VCL_STRING_LIST which uses a
struct instead of varargs.

This means that multiple VCL_STRANDS can be passed to a function,
and string-compare really needed that to not waste workspace.

Add <, >, <=, >= string comparisons while here anyway.

The struct and a const char *[N] array are built on the stack and
they are only valid for the duration of the function call.