varnishtest: add -keepalive to repeat on a single connection
For tests which do not require new connections for repetitions (for
example because of possible error conditions), this reduces run time
and the number of required ephemeral ports.

The latter is a real issue when running many vtcs in parallel which
each run many repetitions of the same test (for example to check for
possible race conditions). When ephemeral ports are exhausted,
seemingly unrelated issues like the following can be observed:

---- c1010 14.0 Failed to open 127.0.0.1 59763: (null)

**** v1    1.8 vsl|          0 CLI             - Wr 300 65 Listen failed
on socket '127.0.0.1:33328': Address already in use
**** v1    1.8 vsl|          0 CLI             - EOF on CLI connection,
worker stops

An argument could be made that UDS does not suffer from the port
exhaustion issue and thus such tests could be migrated to UDS. Yet
also for this case the run time point remains, plus deliberately
testing many iterations on a single connection could have its own
merits.