Add a v_* attribute for deprecated functions
This works with gcc 6.3.0 and clang 3.8.1-24

The test for __GNUC__ is deliberately simple and might not catch all
compilers which would potentially support deprecation marks. While more
specifics could be added, the aim is to raise awareness with developers
and we consider it quite unlikely that anyone does not compile with one
of the main stream compilers at all.