vtest: Implement logexpect alternatives
This was another great idea by Dridi, thank you!

We see cases in VSL where the order of some events can not be (is hard
to) fix. For these cases, we add "all of these have to match, but we do
not care about the order" semantics through "?" in the skip column of
"expect" lines.

Example:

        expect 0 =    Debug             1
        expect ? =    Debug             a
        expect ? =    Debug             b
        expect ? =    Debug             c
        expect 0 =    Debug             2

After a line has matched "1", lines matching a, b or c in any order must
follow and then a line matching "2".

In the log, an "alt" line is emitted whenever an alternative expect is
evaluated.

When an alternative matches, matching will continue with the same
"expect" as before.

Implementation:

The skip_max field of (struct logexp_test) was already used for LE_ANY
if not a non-negative integer.

We add LE_ALT and LE_SEEN. The former denotes an expect with
alternatives (that is, the instruction to look for alternatives after
this expect) and the latter denotes an LE_ALT that has been matched.
Obviously, LE_SEEN is internal and can not result from the logexpect
configuration.

When a match on an LE_ALT fails, we do not advance le->test, but rather
look forward in the expect list for any other altnertives and try them
until we either have a match or run out of alternatives.