vav: Treat a trailing comma as a separator
Let's consider the following VAV strings:

    "foo bar baz"
    "foo,bar,baz"
    " foo bar baz "
    " foo,bar,baz "
    "  foo  bar  baz  "

They are all equivalent because consecutive spaces are considered to
form a single separator. However, consecutive commas aren't:

    "foo,bar,baz"
    "foo,,bar,,baz"

In the example above the first string has 3 arguments while the second
has 5 of them. This behavior was however inconsistent with trailing
commas:

    "foo,bar,baz"
    "foo,bar,baz,"
    "foo,bar,baz,,"

When it comes to trailing commas the first two strings above would
contain 3 arguments, and the last string would contain 4 arguments.

With this change, they respectively contain 3, 4 and 5 arguments.