vnum: Work around uninitialized variable in test code
The value should always be set, because we already assign zero on error,
but it might escape (pun intended) the vigilance of the compiler.

Instead of always making `VNUM_2bytes()` assign its second parameter, it
was decided that fixing the test case would be enough since we already
require error handling.