haproxy: implement sd_notify for mworker mode
With the previous implementation of the haproxy mworker mode, VTest was
waiting for the pidfile to be created to start the tests. From time to
time it was causing issues, because the pidfile could be ready before
haproxy is completely ready. With haproxy 3.1 and the new master-worker
model this is even worse, because the pidfile is created before parsing
the configuration, so haproxy is never ready.

This patch implements the -Ws mode of haproxy, which was initially made
to support the sd_notify API of systemd. This mode sends READY=1 message
over a SOCK_DGRAM unix socket once haproxy is completely ready.
In order to receive the message, VTest binds on a UNIX sockets and
stores the path of this socket in the NOTIFY_SOCKET environment
variable. Once the message is received, VTest will start the tests.