varnishtest: reduce delta to vtest
In preparation of a new attempt to merge varnishtest and vtest, change all
a*.vtc self-tests to invoke varnishtest as vtest and use the vtest magic at the
beginning of the vtc file.

For invocation as vtest, we create links. For the installed binary, this is done
via a hook, but to support varnishtest -i usage, we also need to create the link
in the builddir.