varnishtest: Fix argument processing with vtc_gzip_cmd()
vtc_gzip_cmd() returns 0 for "not a gzip argument", 1 for "gzip argument with
value" and 2 for "same and also add content-encoding". For 2, the argument
vector was increased by 2 also, wrongly skipping an additional argument of av
(and causing an out-of-bounds access, which usually was covered but did surface
while working on an unrelated).